Website with menu
Was unable to eliminate the white space between content and header.
There is also white space on the left and right side of divs. How do I
get rid of it?

Also, used display: inline-block for the divs to get rid of the white
space in between, not sure if this is the right way to do it.